Added better namespace handling
Refactored to allow the library to handle namespaces and inheritance more intelligently.

In order to achieve this, there is the beginnings of an implementation of a method to generate the `jsonld` representation of the OpenActive spec, via `getContext` and `getGraph`. `getGraph` still requires some work before it's finished.

Also implemented a new method, `getFullyQualifiedProperty`, which will take a property name, optional contexts and try to resolve the associated prefix, namespace and alias of the property string.

This work also meant removing `getAlias`, in favour of the new subclassing behaviour.